+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+- fix: [button enabled=false pic=...]で【画像は三等分しない】仕様が三等分していた件
+- fix: [playse]で end_ms 属性エラーチェックの判定ミスでエラーになる件
+- fix: [playse]でループする場合のエラーメッセージで、停止スクリプト表示をしていなかった件
+- docs: タグリファレンスの [graph]インライン画像表示 で属性 x, y は使えないので記述削除
+
+
 ## [1.43.7](https://github.com/famibee/SKYNovel/compare/v1.43.6...v1.43.7) (2023-05-08)